Deleted erronous artist and records that reference it
I'm not sure what happened, but I cannot seem to track down how an artist
name of '2014-12-13 18:26:17' was inserted. The song is 'I WALK THE LINE',
which I assume is the Johnny Cash version. The datetime format is only
identical to the one SQLite/MySQL/(possibly other DBMSes) use and my times
should never go back 2 years. I'm wondering if maybe the screwup was on
the station's website. Still something to be wary about. Here's the
playlist record with the joined tables for complete info. The row is split
on two lines to hopefully make it easier to read if it gets hit with word
wrap.

id   | station_id | station_name       | playlist_url
| song_id | artist_id | song_name
4589 | 5          | The Eagle 103.5 FM |
http://radio.eagle1035.com/songhistory.asp | 2070    | 1085      | I WALK
THE LINE

artist_name         | original_time | formatted_time                  |
posix_timestamp | created
2014-12-13 18:26:17 | 10:50:54 PM   | 17/12/2016 10:50:54 PM UTC-0300 |
1482025854      | 2016-12-18 04:00:19